Revert "add interpolate_like for cpu" (#10645)
Reverts Oneflow-Inc/oneflow#10544

This PR reverts the previous changes that added support for the
interpolate_like operator. During further testing, it was found that the
backward (gradient) implementation of this operator is not correctly
supported, which could lead to incorrect training behavior. We’ll
revisit and reintroduce this feature once the backward implementation is
properly handled and verified.